requirements.txt
tk
opencv-python
pycryptodome
Pillow